move convert.convert to convert.main
