+++ brownie/convert/__init__.py
@@ -1,6 +1,7 @@
 #!/usr/bin/python3
 
-from .convert import (  # NOQA: F401
+from .datatypes import EthAddress, Fixed, Wei  # NOQA: F401
+from .main import (  # NOQA: F401
     to_address,
     to_bool,
     to_bytes,
@@ -9,4 +10,3 @@ from .convert import (  # NOQA: F401
     to_string,
     to_uint,
 )
-from .datatypes import EthAddress, Fixed, Wei  # NOQA: F401
